Behavior spec: Add section about the Create and Offer flows
This new text may seem ambiguous and confusing, and I hope it can
improve, but the purpose of it to remove those verbose details from the
description of how to open a ticket, allowing that text to be much more
concise.

zPlus, you said it's too complicated and will scare away potential
implementors. In the next commit I'll try to simplify it and remove all
the detail now explained in the new section I just added, and let's see
how it goes ^_^