Vocabulary: Add mirror properties (#148)
`mirrors` is meant to represent GitHub and Gitea's concept of mirrors
(respectively called `mirror_url` and `original_url` in their API);
as well as GitLab's pull mirrors (not present in the API, besides a
boolean state).

`mirroredBy` is its inverse property for the sake of completeness.

`mirrorsTo` is meant to represent GitLab's push mirrors (which GitHub
and Gitea don't have); and `mirroredFrom` is its inverse property.

These properties are meant to informational (part of #130),
rather than trigger actions from/to other repositories.

Co-authored-by: Valentin Lorentz <vlorentz@softwareheritage.org>
Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/148
Reviewed-by: fr33domlover <fr33domlover@noreply.codeberg.org>
Reviewed-by: Anthony Wang <xy@noreply.codeberg.org>
Co-authored-by: Val Lorentz <_zn3val@noreply.codeberg.org>
Co-committed-by: Val Lorentz <_zn3val@noreply.codeberg.org>