Fix #173: Move redundant README content to website index.md (#187)
I also renamed COPYING to LICENSE since that seems to be a more typical name for it, and removed a lot of unneeded entries from the .gitignore.

For the README and website, index.md, I mostly replaced the website's contents with text from the README and deleted a lot of old or duplicated information. For instance, I removed the project links at the bottom of the website since there are links to them at the top. Overall, I tried to make the website itself to be non-technical and a good introduction to what ForgeFed is, and then keep technical information in the README of the repo.

This also (hopefully) fixes #132, the missing emojis issue. You can see a live demo of this PR at https://forgefed.exozy.me to verify for yourself if the emojis are working now.

Co-authored-by: Anthony Wang <a@exozy.me>
Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/187
Co-authored-by: Anthony Wang <xy@noreply.codeberg.org>
Co-committed-by: Anthony Wang <xy@noreply.codeberg.org>