Temporarily use archlinux instead of debian for CI until we find a better solution (#192)
The CI currently takes 40 minutes to run and times out when trying to deploy the website, which is a huge blocker. For some reason, using `archlinux` as the container image fixes the slowness problem, but this is only a temporary fix since using `debian` will be more stable in the long run.

Co-authored-by: Anthony Wang <a@exozy.me>
Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/192