Spec: Add support for relying on a previous check of a Grant revocation URI (#208)
(This is just #198 rebased on top of the latest main branch)

Until now, every use of a Grant delegation chain required checking all
the revocation URIs by sending HTTP requests to them. For deeply nested
structures of projects and teams, or for very big projects and teams,
this may cause a lot of network traffic for every resource access, and pressure
on the revocation URIs, and may require very high availability of the
servers hosting those URIs.

This commit allows a Grant to specify a duration, in seconds, such that
the revocation URI check may be skipped if the duration hasn't passed
since the last check of the URI (by the specific actor or by the whole
server). This allows to reduce the impacts described above.

Co-authored-by: Pere Lev <pere@towards.vision>
Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/208