Spec: Use JSON-based authenticity proofs in Revoke activities (#199)
Until now, Revoke activities were allowed to "copy" properties from the
Grants they cancel, because the only embedded signing method on the
Fediverse was based on JSON-LD algorithms and linked data signing (which
are both controversial).

With the object integrity FEP that came in 2022, which is based on plain-JSON
signing and doesn't depend on JSON-LD, we now have a better option.

This PR removes the "copying" of properties from Grant to Revoke,
instead allowing a Revoke to include entire Grant objects, signed using
the method described in the FEP.

Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/199
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Co-authored-by: Pere Lev <pere@towards.vision>
Co-committed-by: Pere Lev <pere@towards.vision>