Spec: Define process of add/remove component to/from project (#210)
This PR defines the sequence of activities for adding a component (e.g. a repo, an issue tracker, etc.) to a project. And the process of removal.

The adding process is implemented in Vervis (it took me a while, was more complicated than I hoped). I haven't yet fully implemented the removal, but it seems simple enough to include in this PR for completeness.

Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/210
Reviewed-by: Aravinth Manivannan <realaravinth@noreply.codeberg.org>
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Co-authored-by: Pere Lev <pere@towards.vision>
Co-committed-by: Pere Lev <pere@towards.vision>