Spec: Issue and PR custom workflows (generalization of issue labels) (#259)
A simple federated schema customization system (based on OpenProject/Redmine), with a Class type meant for representing issue labels.

The system as a whole is probably low priority, but specifically issue labels are an essential feature so I'm adding the whole system. It's definitely possible e.g. in Forgejo to just implement Class fields and ignore everything else until later. But for completeness I'm adding the system as a whole here. It will definitely evolve a lot, it's just a beginning that should be more than enough for now.

Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/259
Co-authored-by: Pere Lev <pere@towards.vision>
Co-committed-by: Pere Lev <pere@towards.vision>