Oversight Teams (#312)
In ForgeFed, teams are standalone entities, each team with its own admin. This setup is meant to support organizations and communities with a non-hierarchical, decentralized, sociocracy-like, hive-like structure. However, in Forgejo and in GitHub, teams don't have admins - instead, there's an "owner" team that does the setup and oversight of the other teams in the organization. Also, even in decentralized organizations, there's often a tech team that might benefit from having wide access to teams, simply to perform technical maintenance tasks and fix issues.

Until now, the ForgeFed solution to such owner teams has been one of these:

1. Have teams delegate access-to-themselves to their child teams (the problem: what if I want to delegate access-to-resources without giving admin access to the team itself? e.g. why should people in the Frontend team be able to freely add and remove people from the Employees team?)
2. Add managers / oversight people as direct collaborators to all teams (problem: UI isn't able to indicate these people are organization managers rather than regular team members)

One possible fix would be to allow a team, when adding a new child, to decide whether to delegate access-to-self. This is good enough, but it makes the organization hierarchy slightly invisible to the human eye: There's a crucial different between child teams that do have that extra access and the child teams that don't.

So, instead, the solution this PR brings is to supplement the parent/child system (where teams DON'T delegate access-to-self ANYMORE from now on) with an oversight/overseen system (where teams delegate *only* access-to-self to their oversight teams). This makes a very clear and visible distinction between the parent-child relationship (meant for teams to share resource access) and the oversight relationship (meant for teams to be fully controlled by other teams).

This also sets the ground for a ForgeFed representation of GH-flavored "organizations": On a typical "forge", every team has exactly one oversight team, and that's the organization's owner team.

Clarification: Even this oversight system still requires mutual consent between the teams. A team cannot force oversight on a team hosted on another server. Just like in all other actor relationships in ForgeFed, mutual consent is established, and one side revoking consent cancels the oversight relationship (therefore the only way to force oversight is to host both teams on the same server and deny the overseen team's ability to revoke consent).
Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/312
Co-authored-by: Pere Lev <pere@towards.vision>
Co-committed-by: Pere Lev <pere@towards.vision>