Spec: SSH Keys and Remote Push (#316)
The federated push mechanism I picked here, after examining the options, is to **embed the Grant URI into the path component of the push URI**. The reason for that is that in the general case it's not practical for a `Repository` actor to successfully maintain a set of all SSH keys of all the users who indirectly have write access to the repo. It might be easy for the direct collaborators, but if a whole tree of teams has access to the repo, and/or the repo is placed under a project hierarchy, this becomes complicated.

So, instead, the Grant URI is provided, allowing the repo to verify authorization in the same way every other federated action is authorized, as well as to identify the actor performing the push and thus trivially verify the SSH keys match.

Note that once an actor pushes, the `Repository` can cache the key, such that further pushes won't require key download. HOWEVER they will still require verifying authorization.

Reviewed-on: https://codeberg.org/ForgeFed/ForgeFed/pulls/316
Co-authored-by: Pere Lev <pere@towards.vision>
Co-committed-by: Pere Lev <pere@towards.vision>